factor browser => common for REPL
