+++ src/vs/workbench/contrib/positronConsole/browser/positronConsoleView.tsx
@@ -21,7 +21,7 @@ import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/la
 import { PositronConsole } from 'vs/workbench/contrib/positronConsole/browser/positronConsole';
 import { ILanguageRuntimeService } from 'vs/workbench/services/languageRuntime/common/languageRuntimeService';
 import { IReactComponentContainer, ISize, PositronReactRenderer } from 'vs/base/browser/positronReactRenderer';
-import { IReplService } from 'vs/workbench/contrib/repl/browser/repl';
+import { IReplService } from 'vs/workbench/contrib/repl/common/repl';
 
 /**
  * PositronConsoleViewPane class.
