+++ src/vs/workbench/contrib/repl/common/replService.ts
@@ -6,8 +6,8 @@ import { Emitter } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ILogService } from 'vs/platform/log/common/log';
 import { ILanguageService } from 'vs/editor/common/languages/language';
-import { ReplInstance } from 'vs/workbench/contrib/repl/browser/replInstance';
-import { ICreateReplOptions, IReplInstance, IReplService } from 'vs/workbench/contrib/repl/browser/repl';
+import { ReplInstance } from 'vs/workbench/contrib/repl/common/replInstance';
+import { ICreateReplOptions, IReplInstance, IReplService } from 'vs/workbench/contrib/repl/common/repl';
 import { formatLanguageRuntime, ILanguageRuntime, ILanguageRuntimeService, RuntimeState } from 'vs/workbench/services/languageRuntime/common/languageRuntimeService';
 
 /**
